+++ tests/bip_173_test_vectors.rs
@@ -60,7 +60,7 @@ macro_rules! check_valid_address_roundtrip {
                 // tested by the test vectors. However when BIP-350 came into effect only witness
                 // version 0 uses bech32 (and this is enforced by encode/decode).
                 if let Ok((hrp, bech32::Fe32::Q, program)) = bech32::segwit::decode($addr) {
-                    let encoded = bech32::segwit::encode_v0(hrp, &program).expect("failed to encode address");
+                    let encoded = bech32::segwit::encode_v0(&hrp, &program).expect("failed to encode address");
                     // The bips specifically say that encoder should output lowercase characters so we uppercase manually.
                     if encoded != $addr {
                         let got = encoded.to_uppercase();
@@ -72,7 +72,7 @@ macro_rules! check_valid_address_roundtrip {
                 let hrp = hrpstring.hrp();
                 let witness_version = hrpstring.witness_version();
 
-                let encoded = hrpstring.byte_iter().bytes_to_fes().with_checksum::<Bech32>(&hrp.into()).with_witness_version(witness_version).chars().collect::<String>();
+                let encoded = hrpstring.byte_iter().bytes_to_fes().with_checksum::<Bech32>(hrp).with_witness_version(witness_version).chars().collect::<String>();
 
                 // The bips specifically say that encoder should output lowercase characters so we uppercase manually.
                 if encoded != $addr {